third_party/blink/web_tests/external/wpt/css/css-values/minmax-percentage-serialize-expected.txt

This is a testharness.js-based test.
[FAIL] 'min(1%, 2%, 3%)' as a computed value should serialize as 'min(1%, 2%, 3%)'.
  assert_equals: 'min(1%, 2%, 3%)' should round-trip exactly in computed values. expected "min(1%, 2%, 3%)" but got "1%"
[FAIL] 'min(3%, 2%, 1%)' as a computed value should serialize as 'min(3%, 2%, 1%)'.
  assert_equals: 'min(3%, 2%, 1%)' should round-trip exactly in computed values. expected "min(3%, 2%, 1%)" but got "1%"
[FAIL] 'min(-1%, 1%)' as a computed value should serialize as 'min(-1%, 1%)'.
  assert_equals: 'min(-1%, 1%)' should round-trip exactly in computed values. expected "min(-1%, 1%)" but got "-1%"
[FAIL] 'min(-0%, 0%)' as a computed value should serialize as 'min(0%, 0%)'.
  assert_equals: 'min(0%, 0%)' should round-trip exactly in computed values. expected "min(0%, 0%)" but got "0%"
[FAIL] 'max(1%, 2%, 3%)' as a computed value should serialize as 'max(1%, 2%, 3%)'.
  assert_equals: 'max(1%, 2%, 3%)' should round-trip exactly in computed values. expected "max(1%, 2%, 3%)" but got "3%"
[FAIL] 'max(3%, 2%, 1%)' as a computed value should serialize as 'max(3%, 2%, 1%)'.
  assert_equals: 'max(3%, 2%, 1%)' should round-trip exactly in computed values. expected "max(3%, 2%, 1%)" but got "3%"
[FAIL] 'max(-1%, 1%)' as a computed value should serialize as 'max(-1%, 1%)'.
  assert_equals: 'max(-1%, 1%)' should round-trip exactly in computed values. expected "max(-1%, 1%)" but got "1%"
[FAIL] 'max(-0%, 0%)' as a computed value should serialize as 'max(0%, 0%)'.
  assert_equals: 'max(0%, 0%)' should round-trip exactly in computed values. expected "max(0%, 0%)" but got "0%"
[FAIL] 'calc(min(1%, 2%) + max(3%, 4%) + 10%)' as a computed value should serialize as 'calc(10% + min(1%, 2%) + max(3%, 4%))'.
  assert_equals: 'calc(10% + min(1%, 2%) + max(3%, 4%))' should round-trip exactly in computed values. expected "calc(10% + min(1%, 2%) + max(3%, 4%))" but got "15%"
Harness: the test ran to completion.
